ai_fashion/requirements.txt
django
djangorestframework
huggingface-hub
Pillow
django-cors-headers
transformers 
torch
diffusers